+++ data-structures/assignment-4/range-sums/splay.h
@@ -1,3 +1,3 @@
 #include "btree.h"
 
-void splay(node **root, int value);
+void splay(node **root, unsigned long long value);
